excercicesDeleet excercicesUpdate coachFindByID

+++ CoachingApp/Implementations/ExerciseManager.cs
@@ -1,5 +1,6 @@
 ﻿using CoachingApp.Interfaces;
 using CoachingApp.Models;
+using Microsoft.EntityFrameworkCore;
 
 namespace CoachingApp.Implementations
 {
@@ -10,21 +11,34 @@ namespace CoachingApp.Implementations
         {
             this.context = _context;
         }
+        public List<Excercise> GetAllExcercises()
+        {
+            var GetCoachs = context.Excercises.ToList();
+            return GetCoachs;
+        }
+
         public Excercise UpdateExcersise(int id, Excercise excersise)
         {
-            var OldExcercice = context.Excercises.Where(e=>e.id==id).SingleOrDefault();
+          
 
-            OldExcercice.coachID = excersise.coachID;
-            OldExcercice.link = excersise.link;
-            OldExcercice.description = excersise.description;
+            var OldExcercice = context.Excercises.FirstOrDefault(s => s.id == id);
 
-            context.SaveChanges();
+            if (OldExcercice != null)
+            {
+                OldExcercice.link = excersise.link;
+                OldExcercice.description = excersise.description;
+
+                context.SaveChanges();
+            }
+
+              
             return OldExcercice;
 
         }
         public Excercise DeleteExcercice(int id)
         {
             var ExcerciceData = context.Excercises.Where(e=>e.id==id).SingleOrDefault();
+            if(ExcerciceData!=null)
             context.Excercises.Remove(ExcerciceData);
 
             context.SaveChanges();
